Filter tasks by priority
This commit introduces the ability to filter tasks by priority in the `TaskController`.

The `getAllTasks` endpoint now accepts an optional query parameter `q` of type `TaskPriority`. If the `q` parameter is provided, the endpoint will return a list of tasks filtered by the specified priority. If `q` is not provided, all tasks will be returned.

The `@PathVariable` annotations in `getTask`, `updateTask`, and `deleteTask` have been updated to explicitly specify the path variable name as "id" using `@PathVariable("id")`.